+++ app.js
@@ -16,7 +16,7 @@ import {
   renderMethodology,
   openDialog,
   prefersReducedMotion
-} from "./ui.js?v=104";
+} from "./ui.js?v=105";
 import { ASPECT_KEYS, ASPECT_META } from "./aspects.js";
 import { t, tp, getLang, setLang } from "./i18n.js";
 import { APP_VERSION } from "./version.js";